% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{glm_is_installed}
\alias{glm_is_installed}
\title{Check whether GLM is installed}
\usage{
glm_is_installed()
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if a GLM executable exists in the per-user data directory
(i.e. \code{glm_install()} has been run successfully).
}
